reconfigured gql connection
