+++ blocks/adventure-content-fragment/_adventure-content-fragment.json
@@ -10,7 +10,6 @@
                         "template": {
                             "name": "Adventure Content Fragment",
                             "model": "adventure-content-fragment",
-                            "path": "/content/dam/wknd-shared/en/adventures/surf-camp-in-costa-rica/surf-camp-costa-rica",
                             "filter": "cf"
                         }
                     }
