+++ component-definition.json
@@ -92,7 +92,6 @@
                 "template": {
                   "name": "Adventure Content Fragment",
                   "model": "adventure-content-fragment",
-                  "path": "/content/dam/wknd-shared/en/adventures/surf-camp-in-costa-rica/surf-camp-costa-rica",
                   "filter": "cf"
                 }
               }
